+++ CHANGELOG.md
@@ -35,6 +35,7 @@
         - internally use ``DIRSTACK`` variable
     - now `~+`, `~-` expansions do not check path existence
 - show stack trace of ignored exceptions within finally/defer block
+- imporve some semantic error messages
 
 #### Builtin